move check package into job
