+++ shardingsphere-scaling/shardingsphere-scaling-core/src/main/java/org/apache/shardingsphere/scaling/core/job/check/DataConsistencyCheckerFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.check;
+package org.apache.shardingsphere.scaling.core.job.check;
 
 import lombok.SneakyThrows;
 import org.apache.shardingsphere.scaling.core.job.ShardingScalingJob;
